fix(ci): substitute DOMAIN_NAME in all nginx config files
The workflow was only substituting ${DOMAIN_NAME} in app.conf, but default.conf
also contains this variable. This caused nginx to fail when trying to serve HTTP
requests on port 80, which was needed for the health checks.

Changed to copy all files from sites-enabled/ and use find to substitute DOMAIN_NAME
in all .conf files.